✨ Move directional focus navigation to a new file
